pyproject.toml
[project]
name = "lsms-agent"
version = "0.1.0"
description = "Natural-language analysis agent over LSMS-ISA survey data (8 African countries)."
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "chainlit>=1.3.0",
    "pandas>=2.2",
    "pyarrow>=17.0",
    "pyreadstat>=1.2.7",
    "python-dotenv>=1.0",
    "ipython>=8.30",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "numpy>=2.0",
    "scipy>=1.13",
    "statsmodels>=0.14",
    "linearmodels>=6.0",
    "pypdf>=5.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ingest", "server"]
